feat: migrate to TypeScript (#110)
<!-- 👋 Hi, thanks for sending a PR to package-json-validator! 💖.
Please fill out all fields below and make sure each item is true and [x]
checked.
Otherwise we may not be able to review your PR. -->

## PR Checklist

- [x] Addresses an existing open issue: fixes #91 
- [x] That issue was marked as [`status: accepting
prs`](https://github.com/JoshuaKGoldberg/package-json-validator/issues?q=is%3Aopen+is%3Aissue+label%3A%22status%3A+accepting+prs%22)
- [x] Steps in
[CONTRIBUTING.md](https://github.com/JoshuaKGoldberg/package-json-validator/blob/main/.github/CONTRIBUTING.md)
were taken

## Overview

This change migrates the project to TypeScript. Beyond providing types
now, as well as an ESM option, the aim was to not change the API in a
breaking way.

In doing this migration, I actually found at least one bug (TypeScript
doing its job!). In `validate`, `getSpecMap` was called, and immediately
after there was an if condition to check if the result that came back
was `false`, which would indicate the requested spec name was bad.
However the if condition checked if `specName` was false, rather than
the result of the `getSpecMap` function. That would have allowed people
to pass in nonsense specNames and throwing an error later when
attempting to access `map`. I've fixed that here.

I did put a TODO in though to consider deprecating the one named PJV
export, in favor of directly exporting the functions and constants. This
would make the API more tree-shakable and (in my opinion) more
straightforward. But that should be a separate change, if implemented.

The other quality of life stuff I adjusted:
- Added TsDoc in more places
- Changed concatenatedstrings to template strings

Note: The GH diff is awful on this. VS Code's diff was much better. It
also seemed to lose track of the `git mv` (even though it got it right
in the individual commits)

![file list showing contents of dist
post-build](https://github.com/user-attachments/assets/24b6a4b3-ff6d-442a-bd3c-518755c82b19)

Closes: #91

---------

Co-authored-by: Josh Goldberg <git@joshuakgoldberg.com>